- Fixed connect_until field being a smallint rather than int in the login table (upgrade with upgrade_svn8630.sql)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@8630 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ sql-files/main.sql
@@ -400,7 +400,7 @@ CREATE TABLE `login` (
   `email` varchar(60) NOT NULL default '',
   `level` tinyint(3) NOT NULL default '0',
   `error_message` smallint(11) unsigned NOT NULL default '0',
-  `connect_until` smallint(11) unsigned NOT NULL default '0',
+  `connect_until` int(11) unsigned NOT NULL default '0',
   `last_ip` varchar(100) NOT NULL default '',
   `memo` smallint(11) unsigned NOT NULL default '0',
   `ban_until` int(11) unsigned NOT NULL default '0',
